refactor(gameMap): seperate city component and sidebar component.
